Solve Euler problem #8

+++ src/main/kotlin/sigma/types/Type.kt
@@ -117,7 +117,7 @@ data class IntLiteralType(
 ) : IntType(), PrimitiveLiteralType {
     companion object {
         fun of(
-            value: Int,
+            value: Long,
         ): IntLiteralType = IntLiteralType(
             value = IntValue(value = value),
         )
